Changing how builtin methods registered
- Instead of one function that adds every funtion to the method table; now
  exposing an add_method function that other files can use to register their
  methods in the method table.
- The method that adds these methods still needs to be called.
- This aims to better encapsulate all the methods of each builtin in their
  respective compilation units.

+++ src/libponyc/codegen/genreference.c
@@ -598,7 +598,7 @@ static LLVMValueRef gen_constant_pointer(compile_t* c, ast_t* ast)
 
   pointer_t* ptr = ast_data(ast);
   pony_assert(ptr != NULL);
-  size_t ptr_size = ptr_get_size(ptr);
+  size_t ptr_size = (size_t)ptr_get_size(ptr);
 
   ast_t** ptr_elements = ptr_get_elements(ptr);
   LLVMValueRef elements[ptr_size];
